MOD04-BotCityMaisPython/desafios/desafios-feitos/botproduto/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
botcity-email-plugin==0.1.5
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
botcity-maestro-sdk==0.5.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
imap-tools==1.7.2
importlib_metadata==8.5.0
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
webdriver-manager==4.0.2
zipp==3.20.2
